Create the tasksync command to automatically synchronise tasks.
This introduces a new command, which is designed to be invoked multiple
times, potentially concurrently, by 'at' from the task warrior on-exit
hook script. It uses mkdir to do some basic exclusivity because flock is
not available on Mac, if it encounters the lock directory it will invoke
itself again backing off 60 seconds.